requirements.txt
torch
transformers
sentencepiece
pyyaml
einops
numpy
tqdm
datasets
pillow
scikit-learn